% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_outliers.R
\name{available_outlier_variables}
\alias{available_outlier_variables}
\title{Liefert die im Datensatz tatsaechlich vorhandenen Outlier-Variablen
in der Standardreihenfolge.}
\usage{
available_outlier_variables(data)
}
\description{
Liefert die im Datensatz tatsaechlich vorhandenen Outlier-Variablen
in der Standardreihenfolge.
}
