% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export.R
\name{export_xlsx}
\alias{export_xlsx}
\title{Exportiert alle Ergebnisse als mehrseitige Excel-Datei}
\usage{
export_xlsx(
  params,
  vt_table = NULL,
  file,
  vt1_time = NA_real_,
  vt2_time = NA_real_
)
}
\arguments{
\item{params}{extract_params() Ergebnis}

\item{vt_table}{VT-Zusammenfassungstabelle (tibble)}

\item{file}{Zielpfad}
}
\description{
Exportiert alle Ergebnisse als mehrseitige Excel-Datei
}
