% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stages.R
\name{prepare_normalized_overview}
\alias{prepare_normalized_overview}
\title{Daten fuer normalisierten Verlauf vorbereiten}
\usage{
prepare_normalized_overview(data)
}
\description{
Daten fuer normalisierten Verlauf vorbereiten
}
