% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export.R
\name{quarto_bin}
\alias{quarto_bin}
\title{Erkennt den Pfad zur quarto-Binary (sucht u.a. in /usr/local/bin)}
\usage{
quarto_bin()
}
\description{
Erkennt den Pfad zur quarto-Binary (sucht u.a. in /usr/local/bin)
}
