% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_outliers.R
\name{rolling_median_mad_time}
\alias{rolling_median_mad_time}
\title{Zeitbasiertes Rolling Median + MAD}
\usage{
rolling_median_mad_time(
  time_sec,
  x,
  window_sec = 60,
  align = c("center", "right")
)
}
\arguments{
\item{time_sec}{numerischer Zeitvektor in Sekunden (sortiert)}

\item{x}{numerischer Wertevektor}

\item{window_sec}{Fensterbreite in Sekunden (Gesamtbreite)}

\item{align}{"center" (Default; ±window_sec/2 um t_i)
oder "right"/"trailing" (Fenster [t_i - window_sec, t_i])}
}
\value{
list(median = ..., mad = ..., n = ...) gleicher Laenge wie x
}
\description{
Fuer jeden Punkt i wird der Median und die MAD ueber alle Punkte j
berechnet, fuer die |time_sec[j] - time_sec[i]| <= window_sec/2.
Variable Atemfrequenz fuehrt also zu variabler Sample-Anzahl je Fenster,
aber konstanter Zeitbreite.
}
