% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stages.R
\name{stage_ranges}
\alias{stage_ranges}
\title{Stufen-Zeitbereiche fuer Plot-Hintergrund (Start/End in Sekunden)}
\usage{
stage_ranges(data, stage_col = "stage", time_col = "time_s")
}
\description{
Stufen-Zeitbereiche fuer Plot-Hintergrund (Start/End in Sekunden)
}
